+++ packages/manager/apps/dedicated/client/app/components/dedicated-cloud/dashboard/tiles/options/descriptions/basic-option/basic-option.component.js
@@ -2,8 +2,8 @@ import template from './basic-option.html';
 
 export default {
   bindings: {
+    eol: '<?',
     optionName: '<',
-    presentationUrl: '<?',
     status: '<',
   },
   name: 'ovhManagerPccDashboardOptionsDescriptionsBasicOption',