+++ packages/manager/apps/telecom/src/app/index.js
@@ -1,13 +1,13 @@
 import 'script-loader!jquery'; // eslint-disable-line
 import 'whatwg-fetch';
-import { useShellClient } from '@ovh-ux/shell';
+import { initShellClient } from '@ovh-ux/shell';
 import { isTopLevelApplication } from '@ovh-ux/manager-config';
 import { defineApplicationVersion } from '@ovh-ux/request-tagger';
 import { getShellClient, setShellClient } from './shell';
 
 defineApplicationVersion(__VERSION__);
 
-useShellClient('telecom')
+initShellClient('telecom')
   .then((client) => {
     if (!isTopLevelApplication()) {
       client.ux.startProgress();