+++ packages/manager/apps/public-cloud/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@4.24.25...@ovh-ux/manager-public-cloud@5.0.0) (2023-06-12)
+
+
+### Features
+
+* **generator:**  add generator vitejs react v1 ([#9186](https://github.com/ovh/manager/issues/9186)) ([c911ee6](https://github.com/ovh/manager/commit/c911ee6168e2803e2022dc0e275f242953ad8255))
+
+
+### BREAKING CHANGES
+
+* **generator:** rename useShellClient into initShellClient
+
+
+
+
+
+## [4.24.25](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@4.24.24...@ovh-ux/manager-public-cloud@4.24.25) (2023-06-12)
+
+**Note:** Version bump only for package @ovh-ux/manager-public-cloud
+
+
+
+
+
+## [4.24.24](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@4.24.23...@ovh-ux/manager-public-cloud@4.24.24) (2023-06-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-public-cloud
+
+
+
+
+
+## [4.24.23](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@4.24.22...@ovh-ux/manager-public-cloud@4.24.23) (2023-05-31)
+
+**Note:** Version bump only for package @ovh-ux/manager-public-cloud
+
+
+
+
+
 ## [4.24.22](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@4.24.21...@ovh-ux/manager-public-cloud@4.24.22) (2023-05-30)
 
 **Note:** Version bump only for package @ovh-ux/manager-public-cloud