+++ packages/manager/modules/pci/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.127.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci@5.127.0...@ovh-ux/manager-pci@5.127.1) (2023-06-12)
+
+
+### Bug Fixes
+
+* **pci.billing:** add ai-notebooks on billing page ([#9382](https://github.com/ovh/manager/issues/9382)) ([63d05c4](https://github.com/ovh/manager/commit/63d05c4af9566f192d45719fb7b4168189a854a7))
+* **public-cloud:** upgrade by using installation instead of renew ([#9410](https://github.com/ovh/manager/issues/9410)) ([440a26b](https://github.com/ovh/manager/commit/440a26b54c6071685eeb98f98b0a653c18b1ffdb))
+
+
+
+
+
+# [5.127.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci@5.126.0...@ovh-ux/manager-pci@5.127.0) (2023-06-07)
+
+
+### Bug Fixes
+
+* **pci:** fix active tab display ([#9359](https://github.com/ovh/manager/issues/9359)) ([c49aacc](https://github.com/ovh/manager/commit/c49aacc0fc050c62e7000f72c202306c21baa9fb))
+* **pci:** fix translation not loaded on user add ([#9392](https://github.com/ovh/manager/issues/9392)) ([3f5af57](https://github.com/ovh/manager/commit/3f5af577bd0c0b15a37a309f1322bc7c981d1033))
+
+
+### Features
+
+* **pci.dataprocessing:** eol metrics ([#9211](https://github.com/ovh/manager/issues/9211)) ([9cbb86f](https://github.com/ovh/manager/commit/9cbb86fc48eb08b2f221c6e88090cc9a5f149e80))
+
+
+
+
+
+# [5.126.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci@5.125.0...@ovh-ux/manager-pci@5.126.0) (2023-05-31)
+
+
+### Bug Fixes
+
+* **pci.data-processing:** fix missing executor numbers ([#9235](https://github.com/ovh/manager/issues/9235)) ([9b249fa](https://github.com/ovh/manager/commit/9b249fadc4f9ca26b8bf103370e3e3c1717edd86))
+* **pci:** disable kubeconfig file if its installing ([#9353](https://github.com/ovh/manager/issues/9353)) ([732f679](https://github.com/ovh/manager/commit/732f679cd61912c1f0bd2172ebcf94f93652bbf0))
+
+
+### Features
+
+* **pci.storages:** display monthly prices on create volume backup view ([#9325](https://github.com/ovh/manager/issues/9325)) ([137a979](https://github.com/ovh/manager/commit/137a97992c0ddce717d88c1d2d1e368abdd366dd))
+
+
+
+
+
 # [5.125.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci@5.124.0...@ovh-ux/manager-pci@5.125.0) (2023-05-30)